Fix comparisons

+++ test/random_decimal64_math.cpp
@@ -281,7 +281,7 @@ void random_division(T lower, T upper)
         if (isinf(res) && isinf(res_int))
         {
         }
-        else if (!BOOST_TEST_EQ(res, res_int))
+        else if (!BOOST_TEST_EQ(static_cast<float>(res), static_cast<float>(res_int)))
         {
             std::cerr << "Val 1: " << val1
                       << "\nDec 1: " << dec1
@@ -318,7 +318,7 @@ void random_mixed_division(T lower, T upper)
         if (isinf(res) && isinf(res_int))
         {
         }
-        else if (!BOOST_TEST_EQ(res, res_int))
+        else if (!BOOST_TEST_EQ(static_cast<float>(res), static_cast<float>(res_int)))
         {
             std::cerr << "Val 1: " << val1
                       << "\nDec 1: " << dec1
